.ai/engine/pyproject.toml
[project]
name = "ai-engine"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "langchain-openai",
  "langgraph",
  "langsmith",
  "pyyaml"
]

[project.scripts]
ai-engine = "cli.main:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
